add timestamp when starting a new log
There's a problem with the CCS811 resetting the CO2 and
tVOC levels after you boot the Rpi.  Not sure why it happens, but
it looks like this:

1621186260,24.26,27.46,3736,4421,2,2,2
1621186320,24.24,27.93,3736,4421,3,3,3
1621186380,24.23,27.86,3676,4290,1,2,2
1621186440,24.21,27.72,3689,4320,3,3,3
time,temp,humid,CO2,tVOC,PM1.0,PM2.5,PM10.0
1621186592,24.23,27.48,405,0,2,2,2
1621186652,24.30,27.65,417,2,3,3,3
1621186712,24.31,27.55,409,1,4,4,4
1621186773,24.29,27.00,400,0,3,3,3
1621186833,24.29,27.29,408,1,3,4,4

By logging the boot time we'll know whether it was a hard
boot or if the script died and was restarted.