reset baseline every 10 minutes for testing
This is really hard for me to figure out. I suspect the "baseline"
value is actually a bitmask, but on my development box it drifts
within 10 minutes.  Fine.  Reset it every 10 minutes if it changes.

On my kitchen-based sensor I have seen this drift to the point where
the values read are meaningless.  On 2021-11-03 I preheated the oven
and baked a pizza for ~12 minutes.  The CO2 and tVOC levels went to
absurdly high levels and stayed.  CO2 approached 4000 and tVOC
approached 5000.  I restarted it, it initialized with the value from
the .conf file, and the CO2/tVOC levels also reset.